remove jss uitext
