+++ src/components/Badge/Badge.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import {oneOf, node} from 'prop-types';
 import {Badge as CoreBadge, BadgeProps as CoreBadgeProps} from 'wix-ui-core/StylableBadge';
 import {withStylable} from 'wix-ui-core/withStylable';
-import {UIText} from '../StylableUIText';
+import {UIText} from '../UIText';
 import {SKIN, TYPE, Type, Skin} from './constants';
 import style from './Badge.st.css';
