--- src/components/Button/Button.e2e.ts
@@ -1,17 +1,0 @@
-import * as eyes from 'eyes.it';
-import {browser} from 'protractor';
-import {getStoryUrl} from 'wix-ui-test-utils/protractor';
-import {buttonTestkitFactory} from 'wix-ui-core/dist/src/testkit/protractor';
-
-describe('Button', () => {
-  const storyUrl = getStoryUrl('Components', 'Button');
-  const dataHook = 'storybook-button';
-
-  beforeEach(() => browser.get(storyUrl));
-
-  eyes.it('should render', () => {
-    const driver = buttonTestkitFactory({dataHook});
-
-    expect(driver.element().isDisplayed()).toBe(true);
-  });
-});
