--- src/components/Button/constants.ts
@@ -1,6 +1,0 @@
-export enum Size {
-    TINY = 'tiny',
-    SMALL = 'small',
-    MEDIUM = 'medium',
-    LARGE = 'large'
-}
