+++ src/components/CounterBadge/CounterBadge.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import {oneOf, node} from 'prop-types';
 import {withStylable} from 'wix-ui-core/withStylable';
 import {Badge as CoreBadge, BadgeProps as CoreBadgeProps} from 'wix-ui-core/StylableBadge';
-import {UIText} from '../StylableUIText';
+import {UIText} from '../UIText';
 import {Skin, SKIN} from './constants';
 import {isIcon, isWide} from './utils';
 import style from './CounterBadge.st.css';
