--- src/components/StylableUIText/index.ts
@@ -1,1 +1,0 @@
-export {UIText} from './UIText';
