+++ src/components/UIText/UIText.e2e.ts
@@ -4,7 +4,7 @@ import {getStoryUrl, waitForVisibilityOf} from 'wix-ui-test-utils/protractor';
 import {uiTextTestkitFactory} from '../../testkit/protractor';
 
 describe('UIText', () => {
-  const storyUrl = getStoryUrl('Components', 'StylableUIText');
+  const storyUrl = getStoryUrl('Components', 'UIText');
 
   beforeEach(() => browser.get(storyUrl));
   eyes.it('should display correct content', () => {
