+++ src/components/UIText/UIText.protractor.driver.ts
