+++ stories/index.ts
@@ -1,6 +1,5 @@
 require('./Autocomplete').story();
 require('./Badge').story();
-require('./Button').story();
 require('./Checkbox').story();
 require('./CounterBadge').story();
 require('./Heading').story();
@@ -9,4 +8,3 @@ require('./StylableToggleSwitch').story();
 require('./ToggleSwitch').story();
 require('./Tooltip').story();
 require('./UIText').story();
-require('./StylableUIText').story();
